add pattern

+++ STRATEGY/src/Context.java
@@ -1,11 +1,11 @@
-public class Context { 
+public class Context {
    private Strategy strategy; 
  
    public Context(Strategy strategy){ 
       this.strategy = strategy; 
    } 
  
-   public int executeStrategy(int num1, int num2){ 
+   public double executeStrategy(double num1, double num2){ 
       return strategy.doOperation(num1, num2); 
    } 
 }
