+++ STRATEGY/src/OperationAdd.java
@@ -1,6 +1,6 @@
 public class OperationAdd implements Strategy{ 
    @Override 
-   public int doOperation(int num1, int num2) { 
-      return num1 + num2; 
+   public double doOperation(double num1, double num2) { 
+      return num1 + num2;
    } 
 }
