+++ STRATEGY/src/Strategy.java
@@ -1,4 +1,4 @@
 
 public interface Strategy {
-	public int doOperation(int num1, int num2);
+	public double doOperation(double num1, double num2);
 }
